refactor: extract settings.rb
