+++ lib/anyway_config.rb
@@ -1,30 +1,13 @@
 # frozen_string_literal: true
 
+require "anyway/version"
+require "anyway/settings"
+require "anyway/config"
+require "anyway/auto_cast"
+require "anyway/env"
+require "anyway/loaders"
+
 module Anyway # :nodoc:
-  require "anyway/version"
-
-  require "anyway/config"
-
-  # Use Settings name to not confuse with Config.
-  #
-  # Settings contain the library-wide configuration.
-  class Settings
-    class << self
-      # Define whether to load data from
-      # *.yml.local (or credentials/local.yml.enc)
-      attr_accessor :use_local_files
-
-      # Return a path to YML config file given the config name
-      attr_accessor :default_config_path
-    end
-
-    # By default, use local files only in development (that's the purpose if the local files)
-    self.use_local_files = (ENV["RACK_ENV"] == "development" || ENV["RAILS_ENV"] == "development")
-
-    # By default, consider configs are stored in the ./config folder
-    self.default_config_path = ->(name) { "./config/#{name}.yml" }
-  end
-
   class << self
     def env
       @env ||= ::Anyway::Env.new
@@ -35,14 +18,10 @@ module Anyway # :nodoc:
     end
   end
 
-  require "anyway/auto_cast"
-  require "anyway/env"
-  require "anyway/loaders"
-
   # Configure default loaders
   loaders.append :yml, Loaders::YAML
   loaders.append :env, Loaders::Env
-
-  require "anyway/rails" if defined?(::Rails::VERSION)
-  require "anyway/testing" if ENV["RACK_ENV"] == "test" || ENV["RAILS_ENV"] == "test"
 end
+
+require "anyway/rails" if defined?(::Rails::VERSION)
+require "anyway/testing" if ENV["RACK_ENV"] == "test" || ENV["RAILS_ENV"] == "test"
